+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - You can now export a story web as a .PNG file from the directory context menu.
 - You can now adjust the size of all custom field editors using the grabber in the lower-right corner.  This size can be set differently for each piece of content.  If you want to change the default size for a particular field, you can still do that in the module settings.
 - Arcs now have a journals tab
+- You can now record a voice clip on character entries.  This is an easy way to quickly pull up what accent/voice you used for each character.
 - You can now export a full setting as either a .MD file or a .ZIP file (the .MD plus all the story web .PNGs)from the directory context menu.  This is handy for reviewing in another interface or for feeding to an LLM to ask it questions about the setting/campaign.
 - Many tables can now be put into "grouped" mode, which allows you to create groups of rows and collapse/expand them.  This is particularly useful for large lists like lore, ideas, and to-do items, but is available other places, too.  There is a new config dialog that lets you specify where you want to use this vs. the normal style.
 - Various settings that turn functionality on/off (ex. show types in tree and use to-do list) now are now reflected immediately in the app window, without needing to refresh the browser.